Restrict screen orientation to portrait
This commit modifies the `AndroidManifest.xml` to lock the application's screen orientation to portrait mode.

The `android:screenOrientation="userPortrait"` attribute has been added to the `<activity>` tag for `MainActivity`. This change is made to simplify development as handling landscape mode is not part of the current assignment and would depend on specific requirements.

A `tools:ignore` attribute has been added to suppress warnings related to `DiscouragedApi` and `LockedOrientationActivity`.